requirements.txt
react==16.13.1
react-bootstrap==1.0.1
react-dom==16.13.1
backcall==0.1.0
decorator==4.4.0
Django==2.2.8
django-cors-headers==3.1.1
django-extensions==2.2.3
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0